+++ InstantCredit/InstantCredit/Credit/Calculate/Response.cs
@@ -5,7 +5,7 @@ namespace InstantCredit.Credit.Calculate;
 public class Response
 {
     public bool CreditIssued { get; set; }
-    public decimal Percent { get; set; }
+    public double Percent { get; set; }
 }
 
 public static class CreditResultExtensions
